Fixing 64-bit compilation warning caused by implicit size_t->int cast

+++ imgui.cpp
@@ -3192,7 +3192,7 @@ static void ClosePopup(ImGuiID id)
     if (!IsPopupOpen(id))
         return;
     ImGuiState& g = *GImGui;
-    ClosePopupToLevel(g.OpenedPopupStack.size() - 1);
+    ClosePopupToLevel((int)g.OpenedPopupStack.size() - 1);
 }
 
 // Close the popup we have begin-ed into.
